implemented infinite loading
