+++ src/utils/getDate.tsx
@@ -11,3 +11,9 @@ export const hoursAgo = (numOfHours, date = new Date()) => {
 
   return hoursAgo;
 };
+
+export const dateToUnix = (_date?: Date) => {
+  const date = _date || new Date();
+
+  return Math.floor(date.getTime() / 1000);
+};
